Move type parsing functions to utils.
